dis/requirements.txt
paho-mqtt==1.6.1
clickhouse-driver==0.2.6
kafka-python==2.0.2
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2 